Fix 'cannot read property 'loadChildren' of undefined' error. This change set fixes the build error 'cannot read property 'loadChildren' of undefined' error by adjusting how I defined my routes.
Routes need to be statically defined and I was statically defining all my routes in each module but then dynamically
concatinating them in the app.routing.ts file. I think this let to lazyloading the modules incorrectly.

To fix the error I removed the app.routing.ts file and pushed the creation of the routes into each modules imports list.
Now when the app.module.ts file imports a sub-module not only will it get the components, services, and directives but
it will also _import_ the routings.

I added the blanket redirect route to the bottom of the app.module.ts file's import list to make sure its the last
route processed when routing requests.