Add RemoteHostListService to provider list
This change set updates the RemoteHost module to export the RemoteHostListService
in the provider list via the forRoot() function so that it is only exported once.

Services are suppose to be singletons but if we just put it in the providers list
via the annotation it would be added each time the module is imported. By using
the statis forRoot() function it will only be added once.