Add feature to pull events from the Remote (GitHub)
This change set adds the requires models, components, and services to pull
events from the given remote (i.e. GitHub) and show them to the user.

Currently it shows the events in much the same way as GitHub does, a later
step will be to stack events related to the same PR.

Note: It expects only one remote registered so it only pulls the first
remote's events. A later step will be to aggergate the events from multiple
remote sites.