Update NPM test script
This change set adds the -sm=false flag to the ng test command called by
NPM test. This resolves some crypted errors when running the tests. I was seeing
failed to send (XHR errors) but what was really happending was that my component
was throwing a TypeError. With this flag I got the actual TypeError and stack trace
in the failed test so I was able to track down the error.