Update feed view to sort by latest activity
This change set updates the feeds view to show the list of activities in
the order of newest to oldest. That is an old PR which gets and update
(i.e. someone comments on it) will be moved to the top of the list
so the most active PR is always at the top and stale PRs are at the bottom