Fix link to Actor
The URL for the actor that comes back in the events api is the _API_ url, that
is the URL to the API to fetch the user's data. I want the _HTML_ url or the
URL I can navigate to in order to see the users profile page.

This change set updates the response parser to get the remote host URL as
a parameter and builds the HTML url for the user (luckally it is just
the sites URL plus the users account name) that way when the user clicks
on the Actors name they are taken to that users profile page on the remote site.