:bug:Fix time text to be right alinged
The text at the right of the activity card shows how long ago the activity took
place. It looks better if it is right aligned since it is on the right side of the
card.

It was _floated_ right but if the text wraps to 2 lines there is a gap as the
wrapped text will start on the left. To fix this this change set changes the
float to a text-align to propertly align the text when wrapped.