Add component to show text if no remotes are set
This change set adds a new component which shows a widget on the view with a
call to action text if no remote hosts have been setup.

The idea is that you would use this widget to help direct a user to add remotes
if they haven't set any up yet.