:bug:Fix hamburger menu
The hamburger menu was not responsive, you click on it and it didn't do anything
This was because I forgot to add the bootstrap JS files to the pipeline nor
had I included jquery (a dependancy of bootstrap).

This change set adds the required JS files to the pipeline so that the hamburger
menu actually works.

Also it updates the colour of the hamburger menu so that the bars are visible
on the dark background.