Refactor remotehost service to be sync
The remotehostlist service extracts the remote hosts from local storage, that
operation is sync but was wrapped in an Observable. This means consumers of
this service needed to subscribe to the observable and store a copy of the
list of remote hosts locally. Each consumer had a 'load remote hosts' function
to hook up the observable but they used the list syncronsly so they called
the load function often. This seemed a waste as the list shoudl be small and
could be held in memory.

To cut down on duplication and cut down the number of times we read from local
storage this change set refactors the service to simply return the list of
remote hosts syncrounsly saving the extra subscribe step. It also holds the
list in memory so that if we've already looked up the list in local storage
we don't look it up again.