Update feed service to persist the feed data
The recived_event API returns recent recived events but older ones already
recived are dropped. Because we might be watching a PR over a long time it
would be nice to keep the older activity data so this change set updates
the feed service to save the feed data it gets to the local storage. Each
time it fetches new recived events from the remote it adds them to the local
cache of activity data.