Update profile pic to be the PR owner
The profile pic shown in the activity feed was of the last user who was active
on the pr. That might be confusing as you might know that the PR is so and so's
but if you were the last person to do something on the PR then your profile pic
will be shown.

This change set updates the PullRequest model to hold the owner of the PR, that
is given via the payload.pull_request.user property, so that we can access the
owner of the PR and always show their profile pic