Update activity card to look better on small/large screens
When the app is viewed on a small screen the thumbnail goes away (good) but the
PR title ended up bleeding past the left edige of the card. Also on large screens
the thumbnail, which is shown, was getting to big making the card kind of tall.

This change set refactors the card so that the PR title does not bleed over the
left edge on small screens and that the card gives more space for the PR title
on large screens; which keeps the thumbnail at a resonable size thus keeping the
card height at a resonable size.