Add ability to remove a PR Activity Card
This change set adds a new ability, the user can now _remove_ a PR activity card
from their feed view. If a PR has been finished (merged/closed) you might want
to remove it from you feed to clean it up a bit.

This change set adds that ability, now their is an X in the top corner of the
card which when clicked will remove the card from the list. If new activity
occures the card will come back, with all history intact. The removal of the
card is done soft so when it comes back you still see all the old activity.

PR Activity cards which have been soft deleted will be reaped (hard deleted)
after two weeks of inactivity