Update base url to be relative
By default the base url is / however when behind a proxy which routes trailing
paths to different locations this can cause an issue.

That is if I setup a proxy that rounts https://mydomain/prview to an upstram
running pr-view at https://pr-view.local I'll be able to get the index.html
however all assets will be relative URLs _which includes the path inserted
by the proxy_ (i.e. https://prview.local/prview/style.css)

To resolve this this change set is updating the base url set in in the
index.html file to be './' (notice the dot). That seems to make things
works without breaking a non-proxy setup