:bug:Change reading state from action to state
The PR event when closing a PR has a payload.action property which reads 'closed'
however it also had a payload.pull_request.state property which also read 'closed'

The issue is that if you comment on a PR, which is an issue event, it has a
payload.action property which reads 'created' because you created a comment;
however, if you comment after a PR is closed this means the app thought that
the PR was open again (i.e. created). Since both PR and Issue payloads have
a state property (payload.pull_request/issue.state) it is better to read
that property thus saving us from thinking a PR was re-opened after being
merged